pyproject.toml
[tool.poetry]
name = "my-bank-app"
version = "0.1.1"
description = ""
authors = ["Guillaume pot <guillaume.pot@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.2.0"
pandas = "^2.2.2"
streamlit = "^1.34.0"
cryptography = "^42.0.7"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
fastapi = "^0.111.0"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
